pyproject.toml
[project]
name = "aethergrid"
version = "0.1.0"
description = ""
authors = [
    {name = "Claude",email = "noreply@anthropic.com"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "weaviate-client (>=4.18.0,<5.0.0)",
    "openai (>=2.8.0,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pymongo (>=4.15.4,<5.0.0)",
    "redis (>=7.0.1,<8.0.0)",
    "fastapi (>=0.121.2,<0.122.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "pydantic (>=2.12.4,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "tiktoken (>=0.12.0,<0.13.0)",
    "aioredis (>=2.0.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.1,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "black (>=25.11.0,<26.0.0)",
    "ruff (>=0.14.5,<0.15.0)",
    "mypy (>=1.18.2,<2.0.0)"
]
